modify code for schema change for rt/nrt rule flags

+++ filter/src/main/java/com/ebay/app/raptor/chocolate/filter/rules/CampaignClickThroughRateRule.java
@@ -39,7 +39,7 @@ public class CampaignClickThroughRateRule extends BaseFilterRule {
    * @return fail weight
    */
   @Override
-  public float test(FilterRequest event) {
+  public int test(FilterRequest event) {
     long campaignId = event.getCampaignId();
     int numerator, denominator;
